Add a LICENSE.txt file
The file is a copy of the CERN-OHL-P v2. Adding it to the tree
is basically required by the license itself, as it ought to be
bundled with the files released under it.

While being here, I have also been asked to comment on the screws
used, so added a sentence about them.

Also added the *-lib-table files since they describe the
project-local libraries for KiCad.

This is supposed to fix issue #5